requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
oss2>=2.17.0
pathlib2>=2.3.0; python_version < "3.4"
